Introduce abstract UniqueObject class for all unqiue objects identifiable by uuid
