+++ dc-model-jackson/src/test/resources/serializedTestObjects/identifiable/alias/UrlAlias.json
@@ -9,9 +9,9 @@
   "targetUuid" : "c31593a4-9620-484b-a4f5-4112731d953b",
   "uuid" : "d1dfbfbf-364e-4983-9d15-c33549145928",
   "website" : {
+    "uuid" : "9b676195-f505-400c-acb2-b7aabb7c62f7",
     "identifiers" : [ ],
     "type" : "ENTITY",
-    "uuid" : "9b676195-f505-400c-acb2-b7aabb7c62f7",
     "entityType" : "WEBSITE",
     "refId" : 0
   }
