+++ dc-model-jackson/src/test/resources/serializedTestObjects/identifiable/entity/work/Item.json
@@ -4,9 +4,9 @@
     "de" : "Zimmer-Gymnastik ohne Geräte : 50 tägliche Übungen für die gesamte Körpermuskulatur, zur Erhaltung der Gesundheit und Förderung der Gewandtheit"
   },
   "previewImage" : {
+    "uuid" : "6bed2ff9-4ad5-4e18-b520-bb9843fe9a73",
     "identifiers" : [ ],
     "type" : "RESOURCE",
-    "uuid" : "6bed2ff9-4ad5-4e18-b520-bb9843fe9a73",
     "fileResourceType" : "IMAGE",
     "readonly" : false,
     "sizeInBytes" : 0,
