+++ dc-model-jackson/src/test/resources/serializedTestObjects/view/RenderingTemplate.json
@@ -1,10 +1,10 @@
 {
+  "uuid" : "2552a82e-8d5c-4bfd-87df-e2f185476885",
   "description" : {
     "en" : "Template description"
   },
   "label" : {
     "en" : "Template label"
   },
-  "name" : "Template name",
-  "uuid" : "2552a82e-8d5c-4bfd-87df-e2f185476885"
+  "name" : "Template name"
 }
