+++ dc-model/src/main/java/de/digitalcollections/model/semantic/Headword.java
@@ -1,8 +1,7 @@
 package de.digitalcollections.model.semantic;
 
-import java.time.LocalDateTime;
+import de.digitalcollections.model.UniqueObject;
 import java.util.Locale;
-import java.util.UUID;
 
 /**
  * See https://en.wikipedia.org/wiki/Headword and
@@ -18,13 +17,10 @@ import java.util.UUID;
  * <p>As it is often interchangeable, we decided for practial reasons to use this class "Headword"
  * also for a lemma.
  */
-public class Headword {
+public class Headword extends UniqueObject {
 
-  private LocalDateTime created;
   private String label;
-  private LocalDateTime lastModified;
   private Locale locale;
-  private UUID uuid;
 
   public Headword() {}
 
@@ -33,43 +29,19 @@ public class Headword {
     this.locale = locale;
   }
 
-  public LocalDateTime getCreated() {
-    return created;
-  }
-
   public String getLabel() {
     return label;
-  }
-
-  public LocalDateTime getLastModified() {
-    return lastModified;
   }
 
   public Locale getLocale() {
     return locale;
   }
 
-  public UUID getUuid() {
-    return uuid;
-  }
-
-  public void setCreated(LocalDateTime created) {
-    this.created = created;
-  }
-
   public void setLabel(String label) {
     this.label = label;
   }
 
-  public void setLastModified(LocalDateTime lastModified) {
-    this.lastModified = lastModified;
-  }
-
   public void setLocale(Locale locale) {
     this.locale = locale;
-  }
-
-  public void setUuid(UUID uuid) {
-    this.uuid = uuid;
   }
 }
