+++ dc-model/src/main/java/de/digitalcollections/model/view/RenderingTemplate.java
@@ -1,19 +1,18 @@
 package de.digitalcollections.model.view;
 
+import de.digitalcollections.model.UniqueObject;
 import de.digitalcollections.model.text.LocalizedText;
-import java.util.UUID;
 
 /**
  * Defines a rendering template to be used in the system implementing this library. It is up to the
  * implementor to add all relevant rendering templates to his system. Each RenderingTemplate can be
  * described by a label, a description and a name.
  */
-public class RenderingTemplate {
+public class RenderingTemplate extends UniqueObject {
 
   private LocalizedText description;
   private LocalizedText label;
   private String name;
-  private UUID uuid;
 
   public LocalizedText getDescription() {
     return description;
@@ -27,10 +26,6 @@ public class RenderingTemplate {
     return name;
   }
 
-  public UUID getUuid() {
-    return uuid;
-  }
-
   public void setDescription(LocalizedText description) {
     this.description = description;
   }
@@ -41,9 +36,5 @@ public class RenderingTemplate {
 
   public void setName(String name) {
     this.name = name;
-  }
-
-  public void setUuid(UUID uuid) {
-    this.uuid = uuid;
   }
 }
